회원 탈퇴 모듈화 적용
