+++ src/api/axiosApi/mypageApi.js
@@ -2,4 +2,5 @@ export const mypage = {
   userDetail: "/api/mypage/userDetail.do",
   userDetailUpdate: "/api/mypage/updateUserInfo.do",
   userPwUpdate: "/api/mypage/updatePwBCrypt.do",
+  withDraw: "/api/mypage/deleteUser.do",
 };
