+++ uploads/ocr_result(25-01-07).txt
@@ -1,0 +1,12 @@
+더이츠푸드
+1월7일 (화) 점심
+흰쌀밥/검은쌀잡곡밥
+부대찌개
+오징어 제육불고기
+순살파닭치킨
+미트볼토마토소스
+스파게티
+크레미에그스크램블
+치컬커리겉절이
+배추김치 겉절이
+양배추샐러드/떡
